requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2==2.9.5
pycodestyle==2.10.0
python-decouple==3.6
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
